refactor(trigger): extract provider running into provider.go
Pure move of the Engine's provider-running surface out of engine.go:
the envresolve.ProviderRunner implementation (Run), the
DenoRuntimeAPI/PythonRuntimeAPI interfaces and their setters, the
shared env resolver accessor (Resolver), and the pre-dispatch env
resolution preflight (preflightEnv). No behavior change; code
relocated verbatim.

Part of #388 (god-file decomposition).
